petitboot: Use linux terminal type
Switching from vt100 to the linux termtype appears to solve the problem
of incorrectly displaying ncurses extended characters used for parts of
the window border. This error can occur, for example, by opening a SOL
session when the Petitboot UI is already drawn:

 Petitboot (v1.5.0-p60e3423)                          8335-GCA 2108ECA
 qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq

  System information
  System configuration
  System status log
  Language
  Rescan devices
  Retrieve config from URL
 *Exit to shell
 qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq
 Enter=accept, e=edit, n=new, x=exit, l=language, g=log, h=help

Switching to 'linux' correctly draws the line borders even when starting
a session late.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>